requirements/ubuntu.txt
--extra-index-url https://download.pytorch.org/whl/cu128
-f https://data.pyg.org/whl/torch-2.6.0+cu128.html

# PyTorch
torch>=2.6.0
torchvision>=0.21.0
torch-scatter>=2.1.2

# Lightning
pytorch-lightning>=2.1.0

# Others
numpy>=1.24.0
python-box>=7.0.0
PyYAML>=6.0
scipy>=1.10.0
matplotlib>=3.5.0
pandas>=2.0.0
opencv-python>=4.8.0
pypose>=0.7.0
networkx>=3.0